cmd/internal/obj: separate more content-addressable symbols by section
DO NOT REVIEW

[Part of a larger DNR change series.]

This ensures that content-addressable symbols headed for go.func.*,
go.funcrel.*, and runtime.gcbits.* don't get combined with other symbols.

This will become important in a subsequent change,
when we replace some relocations with offsets to a base symbol.
When that happens, we need to ensure that such symbols
are in the section we expect them to be in.

Change-Id: Ie8a7a7ccc0de0f69e3893b6093b843c4ce0d70e0