time: fix timer significantly delayed in some cases
Previously, if the service encounters a long GC time or enters the
state that cannot be preempted for a long time(e.g., systemstack),
and it happened other Ps are busy at this time, it will cause timer
significantly delayed.

Actually, we need a mechanism that can quickly identify the timer
that is severely delayed and check it as soon as possible.

Fixes: #45632
Change-Id: I21efcba54ed19a225c98601f889ad8aa253489c0