cmd/compile: add new code coverage instrumentation pass
DO NOT SUBMIT

Introduce a skeleton version of a new code coverage phase/phass to the
compiler. This patch primarily includes the code to walk each function
and identify 'coverable units'; the addition of actual instrumentation
code (counter updates, etc) will appear in a later patch, also the
code that emits the coverage meta-data symbol. The new pass is enabled
only under an option; does not run by default.

Change-Id: I06efdc7ae19af1ecadc9bf5af9f37d5778be19c4