cmd/compile/internal/coverage: add support for counter insertion
DO NOT SUBMIT

Rolls out support for inserting counter updates as part of the
skeleton coverage instrumentation phase. No meta-data generation yet,
and no runtime support, so we're not actually doing anything with the
counter variables.

Each instrumented function will have its own counter symbol (which is
essentially a BSS sym), so that if a function is dead-coded we can
dead-code the counters as well. The counter symbol consists of a
preamble (which will have the ID of the function, the number of
counters, and [eventually] a pointer to the package meta-data symbol),
then an array of uint32 counter locations.

Change-Id: Id87caef86ddb97949af9b20d7d10c54f01cef56f