cmd/compile/internal/coverage: emit meta-data symbol for code coverage
DO NOT SUBMIT

Update the code coverage instrumentation pass to generate and emit a
meta-data symbol. This is a read-only data symbol containing an
encoded representation of the source information for the functions in
the package. This patch also updates the instrumentation code for
function entry basic blocks to store a reference to the package
meta-data symbol into the counter array for the function.

Change-Id: Ia05be0de81b8a4fdbec15125a5d312e28bfbca68