cmd/compile: make sure shapes have proper indexes for sub-instantiation
The computation for determining the shapes to use at the top of
getInstantation was not always creating shapes with the proper indexes.
If an instantiation is being called from another instantiated function,
we cannot just copy the shape types unchanged, because their indexes may
have changed. So, for type args that already shapes, we still call
Shapify() with the correct index.

This does not fully fix the first example in #48645, but does fix the
second example (simplification of the first). There is still another
issue - possible with converting a closure with closure variables to an
interface using a dictionary (and OCONVIDATA).

Updates #48645

Change-Id: Ibb61c6f9a3c317220fb85135ca87eb5ad4dcff9e