cmd/compile: add coverage meta-data decoder
DO NOT SUBMIT

Add a coverage meta-data decoder, which provides APIs for reading
encoded coverage meta-data and expanding it usable form. This package
is intended to be used in the coverage tooling that reads data files
emitted from coverage runs. Along with the new decoding package is a
unit test that runs the encode/decode paths together to check to make
sure that "decode(encode(X)) == X".

Change-Id: I81d27d8da0b2fcfa5039114a6e35a4b463d19b3c