internal/fuzz: prevent the fuzz fn from corrupting data
There is no requirement that fuzz targets not change the
underlying data provided by the fuzzing engine.
In fact, for []byte, it is expected that the underlying
data may be changed by the function under test.

To avoid this, the worker now keeps a scratch slice of
values which will hold a deep copy of the values to be
passed to the fuzz target.

Fixes #48606

Change-Id: I86573ec4f93b6970ccac8b03a7cdb618a2fe2384