runtime,cmd/compile: register meta-data symbols at package init time
DO NOT SUBMIT

When coverage instrumentation is enabled, emit code into the init
function of the package to register the meta-data symbol with the
runtime, so that we can (eventually) provide access to a list of all
meta-data symbols when emitting the coverage data output file.

To facilitate this, add a new runtime routine for registering coverage
meta-data symbols, for use with the new code coverage implementation.
The list of registered meta-data objects can be used (in a future
patch) as part of coverage data file writing.

Change-Id: I510dcef4c155570c65f53245dfe925a958ad1ca0