cmd/compile: improving printing of type names in func/meth instantiations
Created a new types.Type.UniqString() method for printing names of types
in function/method instantiations. (e.g. f[int] or Value[p.Myint].Set())

UniqString generates a completely unique string description for t with
no spaces. It uses full package paths for all non-builtin type names. It
uses semicolon to separate successive fields of a struct and '#' instead
of space to separate field name and type, if necessary.

To avoid the problem of paths that look the same, but are different due
to vendoring, it also uses base.Flag.Cfg.ImportMap[] to do any needed
rewriting of the package path due to vendoring to get the "real" path.

By using UniqString, we make sure function/method instantiations are
completely unique and 1-1 for the specified types, so there is no chance
of improper deduplication. Also, we eliminate spaces, which seems like a
good idea for function/method/type names.

We also use UniqString for naming shape types, so again the shapes are
1-1 with their underlying types, and shape type names have no spaces.

We do not need an analogous function for types2 right now, since we
create all instantiations using types1 types. In the one case where we
need to create an instantiation during types2-to-types1 translation, we
convert the types to types1 first (see (*irgen).instTypeName2).

Change-Id: Iac4748fa0d0d6f89af59bd51076266986daee945