runtime: add GC CPU utilization limiter
This change adds a GC CPU utilization limiter to the GC. It disables
assists to ensure GC CPU utilization remains under 50%. It uses a leaky
bucket mechanism that will only fill if GC CPU utilization exceeds 50%.
Once the bucket begins to overflow, GC assists are limited until the
bucket empties, at the risk of GC overshoot.

The purpose of this limiter is to mitigate GC death spirals, opting to
use more memory instead.

For #48409.

Change-Id: Ia30f914e683e491a00900fd27868446c65e5d3c2