runtime: avoid reentrant locking by cpuprof.add
When locking prof.signalLock in a signal handler, explicitly detect the
case where the current holder is not another signal handler. The other
mechanisms in place pare down the window where a call to cpuprof.add in
a signal handler might have interrupted a call to setcpuprofilerate on
the same thread; this should eliminate such nesting entirely.

Fixes #48782

Change-Id: I08a9d0e3d4eca74324ccf7d582be0352bbeffbdc