cmd/compile: register exit hook to emit code coverage data
DO NOT SUBMIT

When the new code coverage pass is enabled, emit code into the init
function of the "main" package to register the function
'runtime/coverage.onExitHook' to be invoked when the program
terminates. The hook function will be the one that emits coverage data
for the run into the correct spot. This requires adding code to inject
a hidden import of the "runtime/coverage" package when instrumenting
main.

Change-Id: I8b45bb5f60c987654172e2312d796c924c6bc082