cmd/compile: register exit hook to emit code coverage data
DO NOT SUBMIT

When the new code coverage pass is enabled, emit code into the init
function of the "main" package to register the function
'runtime/coverage.onExitHook' to be invoked when the program
terminates. The hook function will be the one that emits coverage data
for the run into the correct spot. This requires adding code to inject
a hidden import of the "runtime/coverage" package when instrumenting
main. Also with this patch is an option to disable the hook calls (but
not the "runtime/coverage" import) for use with "go test -cover",
where the calls to emit the data will be done by _testmain.go
boilerplate.

Change-Id: I8b45bb5f60c987654172e2312d796c924c6bc082