runtime/coverage: runtime routines to emit coverage data
DO NOT SUBMIT

This patch fleshes out the runtime support for emitting coverage data
at the end of a run of an instrumented binary. Data is emitted in the
form of a pair of files, a meta-out-file and counter-data-outfile,
each written to the dir GOCOVERDIR. The meta-out-file is emitted only
if needed; no need to emit again if an existing meta-data file with
the same hash and length is present.

Change-Id: I59d20a4b8c05910c933ee29527972f8e401b1685