runtime,runtime/coverage: runtime routines to emit coverage data
DO NOT SUBMIT

This patch fleshes out the runtime support for emitting coverage data
at the end of a run of an instrumented binary. Data is emitted in the
form of a pair of files, a meta-out-file and counter-data-outfile,
each written to the dir GOCOVERDIR. The meta-out-file is emitted only
if needed; no need to emit again if an existing meta-data file with
the same hash and length is present.

Currently everything is emitted in little-endian form, which for
the counters is somewhat painful in that we need to visit each
counter value and properly encode it.  It might be better to instead
emit counters in native endianity and then teach the tools to decode
properly in the case of an endianity mismatch.

Also worth noting that as of now the meta-file and counter file are
emitted together at the end of a run, but in fact we could just emit
the meta-file write at the get-go (during execution of the init
routine for "main"). The stumbling block there is that we'd need to
have a way to record the final hash of the meta-data so that it could
be also emitted in the counter data file.

Change-Id: I59d20a4b8c05910c933ee29527972f8e401b1685