cmd/go: support new hybrid coverage instrumentation
DO NOT SUBMIT

If GOEXPERIMENT=coverageredesign is in effect, introduce a new
top-level '-cover' option to "go build" to turn on new-style hybrid
code coverage instrumentation. Similarly, use the new instrumentation
for "go test -cover".

The main effects of "-cover" under the hood are to instrument files at
the package level using cmd/cover and to pass additional options to
the compiler when building instrumented packages.

This patch is incomplete. Areas that need attention:

- needs additional test cases
- needs to be reviewed carefully by a "go" cmd expert
- resolve plan going forward for supporting new style
  "go build -cover" for gccgo/gollvm
- what platforms support "go build -cover"?
- what build modes are compatible with "-cover" (-race? -msan? -asan? fuzz?)
- should -cover cause a 'coverage' build tag to be turned on?
- check to make sure addition of -cover to build cache key is correct

Change-Id: Id65621ff6a8c70a30168c1412c2d6f805ff3b9e7