cmd/go: support "-coverage" build mode
DO NOT SUBMIT

Add a new top-level '-coverage' option to "go build", for turning on
compiler-based code coverage instrumentation. At the moment this
mainly passes the right options to the compiler, and forces an import
of 'runtime/coverage' if we're building the main package.

[Note that eventually we will want the options to be named "-cover",
not "-coverage", but doing this rename would create a conflict with
the test flag "-cover"; that will need to be hashed out before we can
do the rename].

This patch is incomplete. Areas that need attention:

- needs test cases
- needs to be reviewed carefully by a "go" cmd expert
- what is the story on supporting "go build -cover" for gccgo/gollvm?
- what platforms support "go build -cover"?
- what build modes are compatible with "-cover" (-race? -msan? -asan? fuzz?)
- should -cover cause a 'coverage' build tag to be turned on?
- check to make sure addition of -cover to build cache key is correct

Change-Id: Id65621ff6a8c70a30168c1412c2d6f805ff3b9e7