cmd/go: support "-coverage" build mode
DO NOT SUBMIT

Add a new top-level '-coverage' option to "go build", for turning on
compiler-based code coverage instrumentation. At the moment this
mainly passes the right options to the compiler, and forces an import
of 'runtime/coverage' if we're building the main package.

This patch is incomplete. Areas that need attention:

- needs test cases
- needs to be reviewed carefully by a "go" cmd expert
- what is the story on supporting "go build -coverage" for gccgo/gollvm?
- what platforms support "go build -coverage"?
- what build modes are compatible with "-coverage" (-race? -msan? -asan?)
- should -coverage cause a 'coverage' build tag to be turned on?
- check to make sure addition of cov to build cache key is correct

Change-Id: Id65621ff6a8c70a30168c1412c2d6f805ff3b9e7