internal/fuzz: pass fuzz inputs over pipe instead of shared memory
DO NOT SUBMIT: this CL is not complete.

In this CL, I was trying to reduce usage of shared memory. Originally,
shared memory made sense when we only supported []byte, but with the
need to marshal / unmarshal inputs, it no longer helps us much. We
only really need it to record information needed to reconstruct an
input that caused a worker to terminate unexpectedly.

This CL passes inputs to the fuzz RPC through pipes (with the rest of
the arguments) instead of shared memory. I measured no significant
performance difference in our benchmarks. I was hoping the code would
be a bit simpler, but it doesn't seem better or worse.

One important part of this change is that if an entry only has a
filename and no loaded data, the worker can load that file
directly. This means the coordinator doesn't need to load corpus files
into memory at all.

For #48163

Change-Id: Ie7a500e61345b481757d66682eb96581e0a00595